+++ examples/start-end.js
@@ -5,7 +5,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import Calendar from 'rc-calendar';
 import DatePicker from 'rc-calendar/src/Picker';
-import zhCn from 'gregorian-calendar/lib/locale/zh_CN'; // spm error
+import zhCn from 'gregorian-calendar/lib/locale/zh_CN';
 import DateTimeFormat from 'gregorian-calendar-format';
 import GregorianCalendar from 'gregorian-calendar';
 import CalendarLocale from 'rc-calendar/src/locale/zh_CN';
@@ -37,6 +37,7 @@ const Picker = React.createClass({
     const calendar = (<Calendar
       locale={CalendarLocale}
       defaultValue={now}
+      formatter={props.showTime ? formatter : dateFormatter}
       timePicker={props.showTime ? timePickerElement : null}
       disabledDate={props.disabledDate}
     />);